+++ sklearn/metrics/cluster/unsupervised.py
@@ -1,7 +1,8 @@
-""" Unsupervised evaluation metrics. """
+"""Unsupervised evaluation metrics."""
 
 # Authors: Robert Layton <robertlayton@gmail.com>
-#
+#          Arnaud Fouchet <foucheta@gmail.com>
+#          Thierry Guillemot <thierry.guillemot.work@gmail.com>
 # License: BSD 3 clause
 
 import numpy as np
@@ -10,6 +11,12 @@ from ...utils import check_random_state
 from ...utils import check_X_y
 from ..pairwise import pairwise_distances
 from ...preprocessing import LabelEncoder
+
+
+def check_number_of_labels(n_labels, n_samples):
+    if not 1 < n_labels < n_samples:
+        raise ValueError("Number of labels is %d. Valid values are 2 "
+                         "to n_samples - 1 (inclusive)" % n_labels)
 
 
 def silhouette_score(X, labels, metric='euclidean', sample_size=None,
@@ -50,12 +57,12 @@ def silhouette_score(X, labels, metric='euclidean', sample_size=None,
         array itself, use ``metric="precomputed"``.
 
     sample_size : int or None
-        The size of the sample to use when computing the Silhouette Coefficient 
-        on a random subset of the data. 
+        The size of the sample to use when computing the Silhouette Coefficient
+        on a random subset of the data.
         If ``sample_size is None``, no sampling is used.
 
     random_state : integer or numpy.RandomState, optional
-        The generator used to randomly select a subset of samples if 
+        The generator used to randomly select a subset of samples if
         ``sample_size is not None``. If an integer is given, it fixes the seed.
         Defaults to the global numpy random number generator.
 
@@ -87,9 +94,7 @@ def silhouette_score(X, labels, metric='euclidean', sample_size=None,
     n_labels = len(le.classes_)
     n_samples = X.shape[0]
 
-    if not 1 < n_labels < n_samples:
-        raise ValueError("Number of labels is %d. Valid values are 2 "
-                         "to n_samples - 1 (inclusive)" % n_labels)
+    check_number_of_labels(n_labels, n_samples)
 
     if sample_size is not None:
         random_state = check_random_state(random_state)
@@ -201,3 +206,58 @@ def silhouette_samples(X, labels, metric='euclidean', **kwds):
     sil_samples = inter_clust_dists - intra_clust_dists
     sil_samples /= np.maximum(intra_clust_dists, inter_clust_dists)
     return sil_samples
+
+
+def calinski_harabaz_score(X, labels):
+    """Compute the Calinski and Harabaz score.
+
+    The score is defined as ratio between the within-cluster dispersion and
+    the between-cluster dispersion. For :math:`K` cluster, the Calinski and
+    Harabaz score is defined as::
+
+
+    With :math:`B_K` the between group dispersion matrix and :math:`W_K`
+    the within-cluster dispersion matrix defined by::
+    the number of samples in the cluster :math:`k` and the mean of the samples
+    in the cluster :math:`k` and the mean of all the samples.
+
+    Parameter
+    ---------
+    X : array-like, shape (n_samples, n_features)
+        List of n_features-dimensional data points. Each row corresponds
+        to a single data point.
+
+    labels : array-like, shape (n_samples,)
+        Predicted labels for each sample.
+
+    Return
+    ------
+    score: float
+        The resulting Calinski-Harabaz score.
+
+    References
+    ----------
+    .. [1] `T. Calinski and J. Harabasz, 1974. "A dendrite method for cluster
+       analysis". Communications in Statistics
+       <http://www.tandfonline.com/doi/abs/10.1080/03610927408827101>`_
+    """
+    X, labels = check_X_y(X, labels)
+    le = LabelEncoder()
+    labels = le.fit_transform(labels)
+
+    n_samples, _ = X.shape
+    n_labels = len(le.classes_)
+
+    check_number_of_labels(n_labels, n_samples)
+
+    extra_disp, intra_disp = 0., 0.
+    mean = np.mean(X, axis=0)
+    for k in range(n_labels):
+        cluster_k = X[labels == k]
+        mean_k = np.mean(cluster_k, axis=0)
+        extra_disp += len(cluster_k) * np.sum((mean_k - mean) ** 2)
+        intra_disp += np.sum((cluster_k - mean_k) ** 2)
+
+    return (1. if intra_disp == 0. else
+            extra_disp * (n_samples - n_labels) /
+            (intra_disp * (n_labels - 1.)))